refactor(core): split the kernel host into mod/publish/load/api
host.rs had grown to 3,500 lines with publish alone at 771 — ~200 of them
the same per-field map_err chain, now one hoisted set() helper — and no
inline tests to anchor a reader. The split follows the file's own section
comments: mod.rs keeps the types, limits and the render/keys/lifecycle
impl; publish.rs the published-table builders (each gated group now a free
build_* function called from its unchanged group() closure, identical keys
and ordering); load.rs the VM construction and declaration readers; api.rs
the injected plugin API and the sandbox scrub. Every external
crate::kernel::host::X path resolves unchanged, and nothing widened beyond
pub(super).
